+++ DualField.m
@@ -395,7 +395,7 @@
 	return YES;
 }
 
-- (unsigned int)lastLengthReplaced {
+- (NSUInteger)lastLengthReplaced {
 	return lastLengthReplaced;
 }
